pulls interfaces and dummy data into their own files
